+++ zerver/lib/emoji.py
@@ -107,8 +107,8 @@ def check_valid_emoji_name(emoji_name: str) -> None:
     raise JsonableError(_("Emoji name is missing"))
 
 
-def get_emoji_url(emoji_file_name: str, realm_id: int) -> str:
-    return upload_backend.get_emoji_url(emoji_file_name, realm_id)
+def get_emoji_url(emoji_file_name: str, realm_id: int, still: bool = False) -> str:
+    return upload_backend.get_emoji_url(emoji_file_name, realm_id, still)
 
 
 def get_emoji_file_name(emoji_file_name: str, emoji_id: int) -> str: